refactor(sequencer): move actors to actor.rs modules
