+++ lez/sequencer/actors/executor/src/actor.rs
@@ -1,0 +1,303 @@
+use common::{block::Block, transaction::LeeTransaction};
+use kameo::{
+    Actor,
+    actor::{ActorRef, WeakActorRef},
+    error::ActorStopReason,
+    mailbox::{MailboxReceiver, Signal},
+    message::{Context, Message},
+};
+use lee_core::{
+    BlockId,
+    account::{Balance, Nonce},
+};
+use log::{info, warn};
+use mempool::MemPoolHandle;
+use sequencer_core::{
+    SequencerCore, TransactionOrigin, block_publisher::BlockPublisherTrait,
+    config::SequencerConfig, task_group::TaskGroup,
+};
+use tokio::select;
+use tokio_util::sync::CancellationToken;
+
+use crate::{
+    Result,
+    error::Error,
+    protocol::{
+        GetAccount, GetAccountBalance, GetAccountNonces, GetAccountReply, GetBlock, GetBlockRange,
+        GetChannelId, GetChannelIdReply, GetLastBlockId, GetProofsAndRoot, GetTransaction,
+        ProduceBlock, Transaction,
+    },
+};
+
+// TODO: Remove `BP` once this part is moved to a separate actor
+pub struct ExecutorActor<BP: BlockPublisherTrait> {
+    sequencer: SequencerCore<BP>,
+    mempool_handle: MemPoolHandle<(TransactionOrigin, LeeTransaction)>,
+
+    // --- TODO: Remove these fields below ---
+    /// Cancelled when the publisher's drive task terminates (e.g. a panicked
+    /// persist sink); no channel events are processed past that point.
+    driver_cancellation: CancellationToken,
+    /// The core's background tasks, taken before the core was shared. This
+    /// handle owns no reference to the core itself, so without these there is
+    /// nothing to wait on: aborting the main loop only starts the teardown.
+    background_tasks: Vec<TaskGroup>,
+}
+
+impl<BP: BlockPublisherTrait> ExecutorActor<BP> {
+    pub async fn new(config: SequencerConfig) -> Self {
+        let (sequencer, mempool_handle) = SequencerCore::<BP>::start_from_config(config).await;
+
+        let driver_cancellation = sequencer.block_publisher().driver_cancellation();
+        let background_tasks = sequencer.background_tasks();
+
+        Self {
+            sequencer,
+            mempool_handle,
+            driver_cancellation,
+            background_tasks,
+        }
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Actor for ExecutorActor<BP> {
+    type Args = Self;
+    type Error = Error;
+
+    async fn on_start(args: Self::Args, _actor_ref: ActorRef<Self>) -> Result<Self> {
+        Ok(args)
+    }
+
+    #[expect(
+        clippy::integer_division_remainder_used,
+        reason = "Generated by select! macro, can't be easily rewritten to avoid this lint"
+    )]
+    async fn next(
+        &mut self,
+        _actor_ref: WeakActorRef<Self>,
+        mailbox_rx: &mut MailboxReceiver<Self>,
+    ) -> Result<Option<Signal<Self>>> {
+        // TODO: Remove this please
+        for task in &self.background_tasks {
+            if task.any_finished() {
+                return Err(Error::BackgroundTaskFinishedUnexpectedly);
+            }
+        }
+
+        select! {
+            signal = mailbox_rx.recv() => {
+                Ok(signal)
+            }
+            () = self.driver_cancellation.cancelled() => {
+                Err(Error::BlockPublisherFinishedUnexpectedly)
+            }
+        }
+    }
+
+    async fn on_stop(
+        &mut self,
+        _actor_ref: WeakActorRef<Self>,
+        _reason: ActorStopReason,
+    ) -> Result<()> {
+        for tasks in &self.background_tasks {
+            tasks.shutdown().await;
+        }
+
+        Ok(())
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<Transaction> for ExecutorActor<BP> {
+    type Reply = Result<()>;
+
+    async fn handle(
+        &mut self,
+        Transaction { transaction }: Transaction,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.mempool_handle
+            .try_push((TransactionOrigin::User, transaction))
+            .map_err(|_err| Error::MempoolIsFull)
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetBlock> for ExecutorActor<BP> {
+    type Reply = Result<Option<Block>>;
+
+    async fn handle(
+        &mut self,
+        GetBlock { block_id }: GetBlock,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.sequencer
+            .block_store()
+            .get_block_at_id(block_id)
+            .map_err(Into::into)
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetBlockRange> for ExecutorActor<BP> {
+    type Reply = Result<Vec<Block>>;
+
+    async fn handle(
+        &mut self,
+        GetBlockRange { range }: GetBlockRange,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        range
+            .map_while(|block_id| {
+                self.sequencer
+                    .block_store()
+                    .get_block_at_id(block_id)
+                    .map_err(Into::into)
+                    .transpose()
+            })
+            .collect::<Result<Vec<_>>>()
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetLastBlockId> for ExecutorActor<BP> {
+    type Reply = Result<BlockId>;
+
+    async fn handle(
+        &mut self,
+        GetLastBlockId: GetLastBlockId,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        Ok(self.sequencer.chain_height())
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetAccountBalance> for ExecutorActor<BP> {
+    type Reply = Balance;
+
+    async fn handle(
+        &mut self,
+        GetAccountBalance { account_id }: GetAccountBalance,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.sequencer
+            .with_state(|state| state.get_account_by_id(account_id).balance)
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetTransaction> for ExecutorActor<BP> {
+    type Reply = Option<(LeeTransaction, BlockId)>;
+
+    async fn handle(
+        &mut self,
+        GetTransaction { tx_hash }: GetTransaction,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.sequencer
+            .block_store()
+            .get_transaction_by_hash(tx_hash)
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetAccountNonces> for ExecutorActor<BP> {
+    type Reply = Vec<Nonce>;
+
+    async fn handle(
+        &mut self,
+        GetAccountNonces { account_ids }: GetAccountNonces,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.sequencer.with_state(|state| {
+            account_ids
+                .into_iter()
+                .map(|account_id| state.get_account_by_id(account_id).nonce)
+                .collect()
+        })
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetProofsAndRoot> for ExecutorActor<BP> {
+    type Reply = (
+        Vec<Option<lee_core::MembershipProof>>,
+        lee_core::CommitmentSetDigest,
+    );
+
+    async fn handle(
+        &mut self,
+        GetProofsAndRoot { commitments }: GetProofsAndRoot,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        self.sequencer.with_state(|state| {
+            let proofs = commitments
+                .iter()
+                .map(|commitment| state.get_proof_for_commitment(commitment))
+                .collect();
+            (proofs, state.commitment_root())
+        })
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetAccount> for ExecutorActor<BP> {
+    type Reply = GetAccountReply;
+
+    async fn handle(
+        &mut self,
+        GetAccount { account_id }: GetAccount,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        GetAccountReply {
+            account: self
+                .sequencer
+                .with_state(|state| state.get_account_by_id(account_id)),
+        }
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<GetChannelId> for ExecutorActor<BP> {
+    type Reply = GetChannelIdReply;
+
+    async fn handle(
+        &mut self,
+        GetChannelId: GetChannelId,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        GetChannelIdReply {
+            channel_id: *self.sequencer.block_publisher().channel_id().as_ref(),
+        }
+    }
+}
+
+impl<BP: BlockPublisherTrait + Send + 'static> Message<ProduceBlock> for ExecutorActor<BP> {
+    type Reply = Result<()>;
+
+    async fn handle(
+        &mut self,
+        ProduceBlock: ProduceBlock,
+        _ctx: &mut Context<Self, Self::Reply>,
+    ) -> Self::Reply {
+        // Only produce on our turn.
+        if !self.sequencer.is_our_turn() {
+            info!("Not our turn to produce a block, skipping");
+            return Ok(());
+        }
+
+        // Never inscribe a second block at a height we already published: the
+        // channel would carry two chains from there and nothing resolves that.
+        // The head rewinds under us when the sdk orphans our own unfinalized
+        // blocks, and recovers once they finalize, so this is a wait.
+        if let Some(high_water) = self.sequencer.rewound_below_published() {
+            warn!(
+                "Skipping turn: head rewound to {} but block {high_water} is already inscribed; \
+                 waiting for the channel to restore it",
+                self.sequencer.next_block_height().saturating_sub(1),
+            );
+            return Ok(());
+        }
+
+        info!("Our turn: collecting transactions from mempool, creating block");
+        let id = self
+            .sequencer
+            .produce_new_block()
+            .await
+            .map_err(Error::BlockProductionFailed)?;
+
+        info!("Block with id {id} created");
+        Ok(())
+    }
+}
