+++ lez/sequencer/actors/rpc_server/src/lib.rs
@@ -1,109 +1,9 @@
 //! RPC Server Actor serves RPC queries and forwards them to Executor.
 
-use std::net::SocketAddr;
+pub use actor::RpcServerActor;
 
-use bytesize::ByteSize;
-use error::Error;
-use jsonrpsee::server::ServerHandle;
-use kameo::{Actor, actor::ActorRef, mailbox::Signal};
-use log::info;
-use sequencer_core::block_publisher::BlockPublisherTrait;
-use sequencer_service_rpc::RpcServer as _;
-use tokio::select;
-
+pub mod actor;
 pub mod error;
 mod service;
 
-const REQUEST_BODY_MAX_SIZE: ByteSize = ByteSize::mib(10);
-
-pub type Result<T> = std::result::Result<T, Error>;
-
-pub struct RpcServerActor {
-    server_handle: Option<ServerHandle>,
-    addr: SocketAddr,
-}
-
-impl RpcServerActor {
-    pub async fn new<BP: BlockPublisherTrait + Send + 'static>(
-        executor_ref: ActorRef<sequencer_executor_actor::ExecutorActor<BP>>,
-        listen_addr: SocketAddr,
-        max_block_size: ByteSize,
-    ) -> Result<Self> {
-        let server = jsonrpsee::server::ServerBuilder::with_config(
-            jsonrpsee::server::ServerConfigBuilder::new()
-                .max_request_body_size(
-                    u32::try_from(REQUEST_BODY_MAX_SIZE.as_u64())
-                        .expect("REQUEST_BODY_MAX_SIZE should be less than u32::MAX"),
-                )
-                .build(),
-        )
-        .build(listen_addr)
-        .await
-        .map_err(Error::RpcServerSetupFailed)?;
-
-        let addr = server
-            .local_addr()
-            .map_err(Error::LocalAddrRetrievingFailed)?;
-
-        info!("Starting RPC Server on {addr}");
-
-        let service = service::Service::new(executor_ref, max_block_size);
-        let server_handle = server.start(service.into_rpc());
-
-        Ok(Self {
-            server_handle: Some(server_handle),
-            addr,
-        })
-    }
-
-    #[must_use]
-    pub const fn addr(&self) -> SocketAddr {
-        self.addr
-    }
-}
-
-impl Actor for RpcServerActor {
-    type Args = Self;
-    type Error = Error;
-
-    async fn on_start(args: Self::Args, _actor_ref: ActorRef<Self>) -> Result<Self> {
-        Ok(args)
-    }
-
-    #[expect(
-        clippy::integer_division_remainder_used,
-        reason = "Generated by select! macro, can't be easily rewritten to avoid this lint"
-    )]
-    async fn next(
-        &mut self,
-        _actor_ref: kameo::prelude::WeakActorRef<Self>,
-        mailbox_rx: &mut kameo::prelude::MailboxReceiver<Self>,
-    ) -> Result<Option<Signal<Self>>> {
-        let handle = self
-            .server_handle
-            .clone()
-            .expect("Server handle should be present while actor is running");
-
-        select! {
-            signal = mailbox_rx.recv() => {
-                Ok(signal)
-            }
-            () = handle.stopped() => {
-                Err(Error::RpcServerStoppedUnexpectedly)
-            }
-        }
-    }
-
-    async fn on_stop(
-        &mut self,
-        _actor_ref: kameo::prelude::WeakActorRef<Self>,
-        _reason: kameo::prelude::ActorStopReason,
-    ) -> Result<()> {
-        if let Some(server_handle) = self.server_handle.take() {
-            server_handle.stop()?;
-            server_handle.stopped().await;
-        }
-
-        Ok(())
-    }
-}
+pub type Result<T> = std::result::Result<T, error::Error>;
